relocate logging logic
