+++ webdev/lib/src/command/daemon_command.dart
@@ -12,6 +12,7 @@ import '../daemon/app_domain.dart';
 import '../daemon/daemon.dart';
 import '../daemon/daemon_domain.dart';
 import '../serve/dev_workflow.dart';
+import '../serve/logging.dart';
 import '../serve/utils.dart';
 import 'configuration.dart';
 import 'shared.dart';
@@ -50,11 +51,10 @@ class DaemonCommand extends Command<int> {
   Future<int> run() async {
     var daemon = Daemon(_stdinCommandStream, _stdoutCommandResponse);
     var daemonDomain = DaemonDomain(daemon);
-    // Override the default logHandler.
-    logHandler = (level, message) {
+    setLogHandler((level, message, {verbose}) {
       daemonDomain.sendEvent(
           'daemon.logMessage', {'level': '$level', 'message': message});
-    };
+    });
     daemon.registerDomain(daemonDomain);
     var configuration = Configuration(launchInChrome: true, debug: true);
     var pubspecLock = await readPubspecLock(configuration);
